# openapi.yaml
openapi: 3.0.3
info:
  title: Defellix API
  description: >
    API Documentation for the Defellix Decentralized Freelancer Trust Platform.
    This API gateway routes requests to underlying microservices: Auth, User, Contract, and Blockchain.
  version: 1.0.0
servers:
  - url: https://api.defellix.com
    description: Production API via Cloudflare Tunnels
  - url: http://localhost:8080
    description: Local development via Nginx API Gateway
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your JWT token in the format **Bearer &lt;token&gt;**
  schemas:
...
    CreateContractRequest:
      type: object
      required: [project_category, project_name, total_amount, client_name, client_email, milestones]
      properties:
        project_category:
          type: string
        project_name:
          type: string
        description:
          type: string
        due_date:
          type: string
          format: date-time
        total_amount:
          type: number
        currency:
          type: string
          default: INR
        prd_file_url:
          type: string
        submission_criteria:
          type: string
          description: Contract-level submission criteria (optional).
          example: "Video URL"
        client_name:
          type: string
        client_company_name:
          type: string
        client_email:
          type: string
          format: email
        client_phone:
          type: string
        terms_and_conditions:
          type: string
        milestones:
          type: array
          items:
            $ref: '#/components/schemas/MilestoneInput'
    UpdateContractRequest:
      type: object
      properties:
        project_category:
          type: string
        project_name:
          type: string
        description:
          type: string
        due_date:
          type: string
          format: date-time
        total_amount:
          type: number
        currency:
          type: string
        prd_file_url:
          type: string
        submission_criteria:
          type: string
          description: Contract-level submission criteria (optional).
          example: "Video URL"
        client_name:
          type: string
        client_company_name:
          type: string
        client_email:
          type: string
          format: email
        client_phone:
          type: string
        terms_and_conditions:
          type: string
        milestones:
          type: array
          items:
            $ref: '#/components/schemas/UpdateMilestoneInput'
    UpdateMilestoneInput:
      type: object
      properties:
        id:
          type: integer
          description: If provided, will update the existing milestone. If omitted, will create a new milestone.
        title:
          type: string
          description: Required if creating a new milestone.
        description:
          type: string
        amount:
          type: number
          description: Required if creating a new milestone.
        due_date:
          type: string
          format: date-time
        is_initial_payment:
          type: boolean
        submission_criteria:
          description: Dynamic JSON, array, or string determining what specific deliverables (urls, files) are needed for this milestone.
          example: "Video URL"
        completion_criteria_tc:
          type: string
          description: Strict Terms & Conditions for milestone approval.
    MilestoneInput:
      type: object
      required: [title, amount]
      properties:
        title:
          type: string
        description:
          type: string
        amount:
          type: number
        due_date:
          type: string
          format: date-time
        is_initial_payment:
          type: boolean
        submission_criteria:
          description: Flexible string or JSON object defining what deliverables are needed.
          example: "Video URL"
        completion_criteria_tc:
          type: string
    ContractResponse:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
        is_revised:
          type: boolean
          description: True if the freelancer has updated the contract after the client requested revisions.
...
    ClientSignRequest:
      type: object
      required: [otp, company_address]
      properties:
        otp:
          type: string
          description: Required 6-digit PIN securely sent to the client's registered email. Use the /send-otp endpoint to generate this code.
        company_address:
          type: string
          description: Required. Can be 'Remote', a full physical address, or a Google Maps URL.
        email:
          type: string
        phone:
          type: string
        company_name:
          type: string
        gst_number:
          type: string
          description: Optional for Indian businesses. Validator to be added in Phase 6.
    SendOTPRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          description: Must exactly match the client_email registered on the contract.

    # --- SUBMISSION MODULE SCHEMAS ---
...
paths:
...
  /api/v1/contracts:
    post:
      tags: [Contract]
      summary: Create a new draft contract
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContractRequest'
      responses:
        '201':
          description: Draft created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractResponse'
    get:
      tags: [Contract]
      summary: List freelancer's contracts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Contracts returned
  /api/v1/contracts/{id}:
    get:
      tags: [Contract]
      summary: Get contract by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Contract details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractResponse'
        '404':
          description: Contract not found
    put:
      tags: [Contract]
      summary: Update a draft or pending contract
      description: Allows a freelancer to edit contract terms and milestones. Only allowed if the contract status is `draft` or `pending` (client requested revisions).
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateContractRequest'
      responses:
        '200':
          description: Contract updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractResponse'
        '403':
          description: Only draft or pending contracts can be edited
        '404':
          description: Contract not found
    delete:
      tags: [Contract]
      summary: Delete a draft contract
      description: Deletes a contract entirely. Only allowed if the status is `draft`.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Contract deleted successfully
        '403':
          description: Only draft contracts can be deleted
        '404':
          description: Contract not found
  /api/v1/contracts/{id}/send:
    post:
      tags: [Contract]
      summary: Send contract to client
      description: Transitions contract from `draft` to `sent`, generates a shareable link token, and triggers an email notification to the client.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Contract sent successfully. Returns the shareable link.
...
  /api/v1/public/contracts/{token}:
    get:
      tags: [Client]
      summary: View contract (Client)
      description: Public endpoint for clients to view the contract using the shareable link token. No auth required.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Contract details returned.
  /api/v1/public/contracts/{token}/send-otp:
    post:
      tags: [Client]
      summary: Send OTP (Client)
      description: Generates a 6-digit PIN and sends it securely to the client's email via SMTP if the email matches the contract's registered address. Required before signing.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendOTPRequest'
      responses:
        '200':
          description: OTP sent successfully
        '400':
          description: Bad request or validation error
        '404':
          description: Contract not found
  /api/v1/public/contracts/{token}/sign:
    post:
      tags: [Client]
      summary: Sign contract (Client)
      description: Client submits their address, demographic information, and the 6-digit OTP code sent to their email. Validating this OTP cryptographically embeds identity trust into the blockchain.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientSignRequest'
      responses:
        '200':
          description: Contract signed and recorded on blockchain.
  /api/v1/public/contracts/{token}/send-for-review:
    post:
      tags: [Client]
      summary: Send contract back for revision (Client)
      description: Allows a client to request changes or add feedback to a `sent` contract before signing it. Changes the status to `pending`.
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
          description: Client view token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [comment]
              properties:
                comment:
                  type: string
                  description: Client feedback/revision request
      responses:
        '200':
          description: Sent for review successfully
        '400':
          description: Bad request or validation error
        '404':
          description: Contract not found
        '409':
          description: Contract is already pending review
...